chore: compile postgres statically to reduce the PLT function call
When building postgres, link all the *.o files directly into the binary
instead of linking libpostgres.so. This approach reduces the overhead of
PLT function calls. Based on our internal performance tests, this change
improves the performance of a 1TB TPC-DS benchmark by 5-8%.

and generate libpostgres.so at the same time, so that other extensions can
write unit tests.