perf: Improve hash agg performance in large-scale data situations
Increase the number of groups processed at a time by agg_fill_hash_table.

When hashagg checks the memory limit, it uses the value of MemoryContextMemAllocated
as the memory usage of the hash table. In fact, this returns the mem_allocated of
the hashkey_mem context. In this case, the memory statistics are inaccurate. For
example, aggstate->hash_mem_limit=4096kb, and the mem_allocated value of the last
round of hashkey_mem_context is 2048kb. When the memory expansion of
hashkey_mem_context is triggered from 2048kb to 4096kb, meta_mem->mem_allocated +
hashkey_mem_context->mem_allocated > aggstate->hash_mem_limit, and
hash_agg_enter_spill_mode is entered. In fact, hashkey_mem_context still has
remaining memory to store hash entries, but we did not use the newly allocated 2MB space.