performance: the agg operator memory should multiply hash_mem_multiplier to reduce the risk of spilling
in cloudberry, our postgres version is 14, and greenplum7's postgres verison is 12.
the size of NumericAggState in cbdb is 144 and it's 128 in greenplum7, In the hash
agg operator, the memory used by a single group is larger than that of gpdb7,we should
refer to the implementation of the postgres kernel and provide a separate
hash_mem_multiplier parameter for the memory used by the hash agg to support dynamic
adjustment to avoid slower performance. The default hash_mem_multiplier refers to
the default value 1.5 of the newer postgres kernel implementation.