Enhancement: change link methd to reduce PLT overhead
During TPC-DS testing, we observed that compiling postgres with
libpostgres.so introduces PTL function call overhead for some functions.
By linking object files (*.o) directly instead, we achieved a 5-8%
performance improvement in the TPC-DS 1TB benchmark.

This commit modifies the postgres linking process to directly link with
the object files rather than the shared library.

Additionally, this update fixes a minor bug: the pax extension has a
dependency on libpostgres.so. Now, when enabling the pax entension, we
check that enable_shared_postgres_backend is set to 'yes' to ensure
proper functionality.